+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ Bug Fixes
 ---------
 * Unable to connect to a cloud cluster using Ubuntu 20.04 (PYTHON-1238)
 * PlainTextAuthProvider fails with unicode chars and Python3 (PYTHON-1241)
+* Graph execution profiles consistency level are not set to LOCAL_QUORUM with a cloud cluster (PYTHON-1240)
 * [GRAPH] Can't write data in a Boolean field using the Fluent API (PYTHON-1239)
 * [GRAPH] Fix elementMap() result deserialization (PYTHON-1233)